requirements.txt
arch
dash
dash-bootstrap-components
feedparser
flask-caching
fredapi
mlflow
numpy
optuna
pandas
plotly
prophet
python-dotenv
redis
requests
scikit-learn
statsmodels
ta
tensorflow
yfinance